+++ internal/uplink/data.go
@@ -197,7 +197,7 @@ func handleCollectedDataUpPackets(ctx common.Context, rxPacket models.RXPacket) 
 	}
 
 	// handle downlink (ACK)
-	time.Sleep(CollectDataDownWait)
+	time.Sleep(common.GetDownlinkDataDelay)
 	if err := downlink.SendDataDownResponse(ctx, ns, rxPacket); err != nil {
 		return fmt.Errorf("handling downlink data for node %s failed: %s", ns.DevEUI, err)
 	}